Extracted the search for incompatible add-ons into separate file/class.
